Create writeppm_test.cpp

+++ HelloMrRay/snippets/writeppm_test.cpp
@@ -13,9 +13,9 @@ int main() {
 	for (int j = height - 1; j >= 0; j--) {
 		for (int i = 0; i < width; i++) {
 			//RGB
-			outFile << int((255 * i) / (width*0.1f)) << " "
-					    << int((255 * j) / (height*0.1f)) << " "
-					    << int(255.f * 0.2f) << "\n";
+			outFile << ((255 * i) / (width*0.1f)) << " "
+				<< ((255 * j) / (height*0.1f)) << " "
+				<< (255 * 0.2f) << "\n";
 		}
 	}
 }
